% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_lines}
\alias{get_lines}
\alias{get_text}
\title{FUNCTION_TITLE}
\usage{
get_lines(x)

get_text(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
