% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_exported.R
\name{is_exported}
\alias{is_exported}
\title{FUNCTION_TITLE}
\usage{
is_exported(funs, pkg)
}
\arguments{
\item{funs}{PARAM_DESCRIPTION}

\item{pkg}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
