% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship.R
\name{relationship}
\alias{relationship}
\title{FUNCTION_TITLE}
\usage{
relationship(x, parent = NULL, child = NULL)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{parent}{PARAM_DESCRIPTION, Default: NULL}

\item{child}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
