% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repack.R
\name{repack}
\alias{repack}
\title{FUNCTION_TITLE}
\usage{
repack(x, dir_out = NULL, single_file = TRUE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{dir_out}{PARAM_DESCRIPTION, Default: NULL}

\item{single_file}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
