% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{unpack}
\alias{unpack}
\title{FUNCTION_TITLE}
\usage{
unpack(path = "R", ns = NULL, warn = TRUE)
}
\arguments{
\item{path}{PARAM_DESCRIPTION, Default: 'R'}

\item{ns}{PARAM_DESCRIPTION, Default: NULL}

\item{warn}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
